dev-util/gammaray: Re-arrange probe and switch to Qt6
There's some problems with how Qt6 probes work in the Qt5 client, but
the client should really be built for Qt6 by now. This replaces the Qt6
probe with a Qt5 one (because the Qt6 probe will be built by the main
GammaRay package now.)

The dependencies still need to be fixed, will be fixed in the next
commit.